+++ system/modules/isotope/dca/tl_iso_producttypes.php
@@ -392,7 +392,7 @@ class tl_iso_producttypes extends Backend
 	 */
 	public function getListTemplates(DataContainer $dc)
 	{
-		return $this->getTemplateGroup('iso_list_');
+		return IsotopeBackend::getTemplates('iso_list_');
 	}
 
 
@@ -403,60 +403,7 @@ class tl_iso_producttypes extends Backend
 	 */
 	public function getReaderTemplates(DataContainer $dc)
 	{
-		return $this->getTemplateGroup('iso_reader_');
-	}
-
-
-	/**
-	 * List template from all themes, show theme name
-	 * Based on Controller::getTemplateGroup from Contao 2.9.3
-	 */
-	protected function getTemplateGroup($strPrefix, $intTheme=0)
-	{
-		$arrThemes = array();
-		$arrTemplates = array();
-		$arrFolders = array();
-
-		// Add the templates root directory
-		$arrFolders[] = TL_ROOT . '/templates';
-
-		// Add theme templates folder
-		$objTheme = $this->Database->execute("SELECT name, templates FROM tl_theme" . ($intPid>0 ? " WHERE id=$intPid" : ''));
-		while( $objTheme->next() )
-		{
-			if ($objTheme->templates != '' && is_dir(TL_ROOT .'/'. $objTheme->templates))
-			{
-				$arrFolders[] = TL_ROOT .'/'. $objTheme->templates;
-				$arrThemes[TL_ROOT .'/'. $objTheme->templates] = $objTheme->name;
-			}
-		}
-
-		// Add the module templates folders if they exist
-		foreach ($this->Config->getActiveModules() as $strModule)
-		{
-			$strFolder = TL_ROOT . '/system/modules/' . $strModule . '/templates';
-
-			if (is_dir($strFolder))
-			{
-				$arrFolders[] = $strFolder;
-			}
-		}
-
-		// Find all matching templates
-		foreach ($arrFolders as $strFolder)
-		{
-			$arrFiles = preg_grep('/^' . preg_quote($strPrefix, '/') . '/i',  scan($strFolder));
-
-			foreach ($arrFiles as $strTemplate)
-			{
-				$strName = basename($strTemplate);
-				$strName = substr($strName, 0, strrpos($strName, '.'));
-				$arrTemplates[$strName] = $strName . (isset($arrThemes[$strFolder]) ? ' ['.$arrThemes[$strFolder].']' : '');
-			}
-		}
-
-		natcasesort($arrTemplates);
-		return $arrTemplates;
+		return IsotopeBackend::getTemplates('iso_reader_');
 	}